api-client/requirements.txt
requests==2.22.0
pandas==1.5.2
python-dotenv==0.21.0